+++ pkg/services/users/users.go
@@ -6,30 +6,39 @@ import (
 	"github.com/iota-agency/iota-erp/sdk/composables"
 	"github.com/iota-agency/iota-erp/sdk/graphql/helpers"
 	"github.com/iota-agency/iota-erp/sdk/service"
-	"gorm.io/gorm"
 )
 
-func NewService(db *gorm.DB) *Service {
-	return &Service{db: db}
+func New() *Service {
+	return &Service{}
 }
 
 type Service struct {
-	db *gorm.DB
 }
 
-func (s *Service) useTx(ctx context.Context) *gorm.DB {
-	if tx, ok := composables.UseTx(ctx); ok {
-		return tx
-	} else {
-		return s.db
+func (s *Service) Get(ctx context.Context, params *service.GetParams[int64]) (*models.User, error) {
+	tx, ok := composables.UseTx(ctx)
+	if !ok {
+		return nil, service.ErrNoTx
 	}
+	q := tx
+	for _, join := range params.Joins {
+		q = q.Joins(join)
+	}
+	user := &models.User{}
+	if err := q.First(user, params.Id).Error; err != nil {
+		return nil, err
+	}
+	return user, nil
 }
 
 func (s *Service) GetAll(ctx context.Context, params *service.FindParams) ([]*models.User, error) {
 	if params == nil {
 		params = &service.FindParams{}
 	}
-	tx := s.useTx(ctx)
+	tx, ok := composables.UseTx(ctx)
+	if !ok {
+		return nil, service.ErrNoTx
+	}
 	q := tx.Offset(params.Offset)
 	if params.Limit > 0 {
 		q = q.Limit(params.Limit)
@@ -55,7 +64,10 @@ func (s *Service) GetAll(ctx context.Context, params *service.FindParams) ([]*mo
 }
 
 func (s *Service) Count(ctx context.Context) (int64, error) {
-	tx := s.useTx(ctx)
+	tx, ok := composables.UseTx(ctx)
+	if !ok {
+		return 0, service.ErrNoTx
+	}
 	var count int64
 	if err := tx.Model(&models.User{}).Count(&count).Error; err != nil {
 		return 0, err
@@ -64,26 +76,29 @@ func (s *Service) Count(ctx context.Context) (int64, error) {
 }
 
 func (s *Service) Create(ctx context.Context, user *models.User) error {
-	tx := s.useTx(ctx)
+	tx, ok := composables.UseTx(ctx)
+	if !ok {
+		return service.ErrNoTx
+	}
 	if err := tx.Create(user).Error; err != nil {
 		return err
 	}
 	return nil
 }
 
-func (s *Service) Update(ctx context.Context, id int64, user *models.User) error {
-	tx := s.useTx(ctx)
-	if err := tx.First(user, id).Error; err != nil {
-		return err
+func (s *Service) Update(ctx context.Context, user *models.User) error {
+	tx, ok := composables.UseTx(ctx)
+	if !ok {
+		return service.ErrNoTx
 	}
-	if err := tx.Save(user).Error; err != nil {
-		return err
-	}
-	return nil
+	return tx.Save(user).Error
 }
 
 func (s *Service) Delete(ctx context.Context, id int64) error {
-	tx := s.useTx(ctx)
+	tx, ok := composables.UseTx(ctx)
+	if !ok {
+		return service.ErrNoTx
+	}
 	if err := tx.Delete(&models.User{}, id).Error; err != nil {
 		return err
 	}
